+++ src/on-posix/hw_specific.h
@@ -47,7 +47,7 @@ extern void threadStartupWrapper(Thread*);
 long* hwInitContext(long* stack, void* object);
 
 /** Node number if a network system **/
-extern long getNodeNumber();
+extern int32_t getNodeNumber();
 
 /********************************************************/
